Combined lib/web.py + lib/spell.py. Added file tracking to Pluggable
In order to make mocking easier (and it just made more sense), I merged
`lib/web.py` into `lib/spell.py`. The new class is now `BaseSpell`
(instead of `spell`)

The testing framework (see next commit) needs to know the directory that
the class belongs to so that it can find the `test_data` directory so
that the tests can be ran off-line. I was having a problem where file
names kept getting screwed up so they are now being cached inside of the
`__metaclass__`.